Byte is actually serialized with Int

+++ example/src/main/scala/example/stockMarket/conditionActionRule.scala
@@ -3,8 +3,8 @@ package stockMarket
 
 class conditionActionRule {
     var strength: Double = 0
-    var label: Byte = 0
-    def eval(stockPrice: Double, marketState: List[Byte]): Byte = ???
+    var label: Int = 0
+    def eval(stockPrice: Double, marketState: List[Int]): Int = ???
     def feedback(success: Boolean): Unit = {
         if (success){
             strength += 1
@@ -18,7 +18,7 @@ class conditionActionRule {
 // Buy if the dividend increases and sell if the dividend decreases
 class rule1(wealth: WealthManagement) extends conditionActionRule {
     label = 1
-    override def eval(stockPrice: Double, marketState: List[Byte]): Byte ={
+    override def eval(stockPrice: Double, marketState: List[Int]): Int ={
         if (marketState(0)==1 && stockPrice < wealth.cash){
             wealth.buyStock(stockPrice)
             1
@@ -35,7 +35,7 @@ class rule1(wealth: WealthManagement) extends conditionActionRule {
 // If both conditions are met, prioritize sell
 class rule2(wealth: WealthManagement) extends conditionActionRule {
     label = 2
-    override def eval(stockPrice: Double, marketState: List[Byte]): Byte ={
+    override def eval(stockPrice: Double, marketState: List[Int]): Int ={
         if (marketState(1) == 1 && wealth.shares >= 1){
             wealth.sellStock(stockPrice)
             2
@@ -51,7 +51,7 @@ class rule2(wealth: WealthManagement) extends conditionActionRule {
 // Buy if 10-day average decreases and sell if 10-day average increases
 class rule3(wealth: WealthManagement) extends conditionActionRule {
     label = 3
-    override def eval(stockPrice: Double, marketState: List[Byte]): Byte ={
+    override def eval(stockPrice: Double, marketState: List[Int]): Int ={
         if (marketState(1) == 2 && stockPrice < wealth.cash){
             wealth.buyStock(stockPrice)
             1
@@ -67,7 +67,7 @@ class rule3(wealth: WealthManagement) extends conditionActionRule {
 // Random buy and sell
 class rule4(wealth: WealthManagement) extends conditionActionRule {
     label = 4
-    override def eval(stockPrice: Double, marketState: List[Byte]): Byte ={
+    override def eval(stockPrice: Double, marketState: List[Int]): Int ={
         if (Random.nextBoolean){
             if (stockPrice < wealth.cash) {
                 wealth.buyStock(stockPrice)
@@ -90,7 +90,7 @@ class rule4(wealth: WealthManagement) extends conditionActionRule {
 // If both conditions are met, prioritize sell
 class rule5(wealth: WealthManagement) extends conditionActionRule {
     label = 5
-    override def eval(stockPrice: Double, marketState: List[Byte]): Byte ={
+    override def eval(stockPrice: Double, marketState: List[Int]): Int ={
         if (marketState(2) == 1 && wealth.shares >= 1){
             wealth.sellStock(stockPrice)
             2
